// src/openapi3/aether/2.1.0/models/site-device.ts
// Code generated by openapi-gen. DO NOT EDIT.
/* tslint:disable */
/* eslint-disable */
import { SiteDeviceState } from './site-device-state';

/**
 * List of devices (single)
 */
export interface SiteDevice {

  /**
   * long description field
   */
  description?: string;

  /**
   * ID for this device.
   */
  'device-id': string;

  /**
   * display name to use in GUI or CLI
   */
  'display-name'?: string;

  /**
   * IMEI for this device
   */
  imei?: string;

  /**
   * Link to simcard
   */
  'sim-card'?: string;
  state?: SiteDeviceState;
}